Rewrite init of network with matrixes + add tests

+++ src/matrix.cpp
@@ -4,20 +4,20 @@
 #include <omp.h>
 
 Matrix Matrix::operator*(Matrix rightMatrix) {
-    int rowsRight = rightMatrix.getRows();
-    int columnsRight = rightMatrix.getColumns();
+    size_t rowsRight = rightMatrix.getRows();
+    size_t columnsRight = rightMatrix.getColumns();
 
     if (columns != rowsRight)
         throw std::invalid_argument("Invalid matrix sizes while multiplying");
 
     std::vector<Scalar> newData;
-    newData.resize(static_cast<size_t>(rows * columnsRight));
+    newData.resize(rows * columnsRight);
     Matrix newMatrix(rows, columnsRight, newData);
 
     #pragma omp parallel for
-    for (int i = 0; i < rows; ++i) 
-    for (int k = 0; k < columns; ++k) 
-    for (int j = 0; j < columnsRight; ++j) 
+    for (int i = 0; static_cast<size_t>(i) < rows; ++i) 
+    for (int k = 0; static_cast<size_t>(k) < columns; ++k) 
+    for (int j = 0; static_cast<size_t>(j) < columnsRight; ++j) 
         newMatrix(i, j) += (*this)(i, k) * rightMatrix(k, j);
 
     return newMatrix;
@@ -28,7 +28,7 @@ void Matrix::randomize() {
     std::mt19937 gen(rd());
     std::normal_distribution<Scalar> dist(0.0, 1.0);
 
-    for (size_t i = 0; i < static_cast<size_t>(rows * columns); ++i) {
+    for (size_t i = 0; i < rows * columns; ++i) {
         dataRow[i] = dist(gen);
     }
 }
